requirements.txt
Django==2.1.2
djangorestframework==3.9.0
djangorestframework-jwt==1.11.0
django-rest-auth==0.9.3
